[#52] refactor: sendersinfo hook 분리
